requirements.txt
requests == 2.10.0
Pillow == 3.2.0
